Changed ConvertTo-Function to include cmdlet names
In the interest of continuing to be able to automatically generate all
the code from scratch I moved the cmdlet names into a hash for easy
identification.

Any further cmdlet names should be added to the
ServiceToCmdletMappings.ps1 file. This will allow us to continue to
easily make changes quickly on a massive / entire code base scale rather
than needing to manually modify any of the resultant code.